# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Building:
    properties:
      Id:
        type: integer
      buildingName:
        type: string
      city:
        type: string
      creationDate:
        type: string
      creatorId:
        type: integer
      region:
        type: string
      shortName:
        type: string
    type: object
  model.BuildingList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Building'
        type: array
    type: object
  model.FailureMsg:
    properties:
      error:
        type: string
    type: object
  model.MachineRole:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      description:
        type: string
      machineRoleName:
        type: string
    type: object
  model.MachineRoleList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.MachineRole'
        type: array
    type: object
...
  model.SuccessMsg:
    properties:
      message:
        type: string
    type: object
...
  model.Vendor:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      vendorName:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    name: Gary Greene
    url: https://github.com/greeneg/allocatord
  description: An API for managing OS imaging
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Allocator Daemon
  version: 0.1.5
paths:
  /building:
    post:
      consumes:
      - application/json
      description: Add a new building
      parameters:
      - description: Building data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/model.Building'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register building
      tags:
      - buildings
  /building/{buildingId}:
    delete:
      consumes:
      - application/json
      description: Delete a building by Id
      parameters:
      - description: Building Id
        in: path
        name: buildingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete building
      tags:
      - buildings
    patch:
      description: Update a building by its Id
      parameters:
      - description: Building ID
        in: path
        name: buildingId
        required: true
        type: integer
      - description: Building data
        in: body
        name: buildingData
        required: true
        schema:
          $ref: '#/definitions/model.Building'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Update a building by its Id
      tags:
      - buildings
  /building/byId/{buildingId}:
    get:
      description: Retrieve a building by its Id
      parameters:
      - description: Building ID
        in: path
        name: buildingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Building'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a building by its Id
      tags:
      - buildings
  /building/byShortName/{buildingShortName}:
    get:
      description: Retrieve a building by its abbreviated name
      parameters:
      - description: Building abbreviation
        in: path
        name: buildingShortName
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Building'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a building by its abbreviated name
      tags:
      - buildings
  /buildings:
    get:
      description: Retrieve list of all building objects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BuildingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all building objects
      tags:
      - buildings
  /machineRole:
    post:
      consumes:
      - application/json
      description: Add a new machine role
      parameters:
      - description: Machine Role data
        in: body
        name: machineRole
        required: true
        schema:
          $ref: '#/definitions/model.Vendor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register machine role
      tags:
      - machine-roles
  /machineRole/{machineRoleId}:
    delete:
      consumes:
      - application/json
      description: Delete a machine role by Id
      parameters:
      - description: Machine Role Id
        in: path
        name: machineRoleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete machine role
      tags:
      - machine-roles
    patch:
      description: Update a machine role by its Id
      parameters:
      - description: Machine Role ID
        in: path
        name: machineRoleId
        required: true
        type: integer
      - description: Machine Role data
        in: body
        name: machineRoleData
        required: true
        schema:
          $ref: '#/definitions/model.MachineRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Update a machine role by its Id
      tags:
      - machine-roles
  /machineRole/byId/{ouId}:
    get:
      description: Retrieve a machine role by its Id
      parameters:
      - description: Machine Role ID
        in: path
        name: machineRoleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MachineRole'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a machine role by its Id
      tags:
      - machine-roles
  /machineRoles:
    get:
      description: Retrieve list of all machine roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MachineRoleList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all machine roles
      tags:
      - machine-roles
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"